Use deleteTreeItemImpl to fix #668
