+++ src/commands/deleteStaticWebApp/IDeleteWizardContext.ts
@@ -3,12 +3,12 @@
 *  Licensed under the MIT License. See License.txt in the project root for license information.
 *--------------------------------------------------------------------------------------------*/
 
-import { AzExtTreeItem, ExecuteActivityContext, IActionContext } from "@microsoft/vscode-azext-utils";
+import { ExecuteActivityContext, IActionContext } from "@microsoft/vscode-azext-utils";
 import { ISubscriptionContext } from "vscode-azureextensiondev";
 import { ResolvedStaticWebAppTreeItem } from "../../tree/StaticWebAppTreeItem";
 
 export interface IDeleteWizardContext extends IActionContext, ExecuteActivityContext {
-    node?: ResolvedStaticWebAppTreeItem & AzExtTreeItem;
+    node?: ResolvedStaticWebAppTreeItem;
     resourceGroupToDelete?: string;
     subscription: ISubscriptionContext;
 }
