+++ src/commands/deleteStaticWebApp/deleteStaticWebApp.ts
@@ -3,34 +3,14 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { AzExtTreeItem, AzureWizard, IActionContext } from '@microsoft/vscode-azext-utils';
+import { AzExtTreeItem, IActionContext } from '@microsoft/vscode-azext-utils';
 import { ext } from '../../extensionVariables';
 import { ResolvedStaticWebAppTreeItem, StaticWebAppTreeItem } from '../../tree/StaticWebAppTreeItem';
-import { createActivityContext } from '../../utils/activityUtils';
-import { localize } from '../../utils/localize';
-import { ConfirmDeleteStep } from './ConfirmDeleteStep';
-import { DeleteResourceGroupStep } from './DeleteResourceGroupStep';
-import { IDeleteWizardContext } from './IDeleteWizardContext';
-import { StaticWebAppDeleteStep } from './StaticWebAppDeleteStep';
 
 export async function deleteStaticWebApp(context: IActionContext, node?: ResolvedStaticWebAppTreeItem & AzExtTreeItem): Promise<void> {
     if (!node) {
-        node = await ext.rgApi.tree.showTreeItemPicker<ResolvedStaticWebAppTreeItem & AzExtTreeItem>(new RegExp(StaticWebAppTreeItem.contextValue), { ...context, suppressCreatePick: true });
+        node = await ext.rgApi.appResourceTree.showTreeItemPicker<ResolvedStaticWebAppTreeItem & AzExtTreeItem>(new RegExp(StaticWebAppTreeItem.contextValue), { ...context, suppressCreatePick: true });
     }
 
-    const wizardContext: IDeleteWizardContext = {
-        ...context,
-        node,
-        subscription: node.subscription,
-        ...(await createActivityContext())
-    };
-
-    const wizard = new AzureWizard<IDeleteWizardContext>(wizardContext, {
-        title: localize('deleteSwa', 'Delete Static Web App "{0}"', node.name),
-        promptSteps: [new ConfirmDeleteStep()],
-        executeSteps: [new StaticWebAppDeleteStep(), new DeleteResourceGroupStep()]
-    });
-
-    await wizard.prompt();
-    await wizard.execute();
+    await node.deleteTreeItem(context);
 }
